Implemented the atof fix regarding starting zeroes within the decimal in float and double values; Standalone properties / values are now subjected to check_range tests; Fixed type integrity loss on ambiant_ratio formerly being a float, now double;

+++ includes/scene.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   scene.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: anle-pag <anle-pag@student.42.fr>          +#+  +:+       +#+        */
+/*   By: epinaud <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/15 20:10:12 by anle-pag          #+#    #+#             */
-/*   Updated: 2025/07/15 16:47:45 by anle-pag         ###   ########.fr       */
+/*   Updated: 2025/11/28 01:56:55 by epinaud          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ typedef struct s_scene
 	t_list		*objects;		//list of t_object*
 	t_list		*lights;		//list of t_light*
 	t_color		ambient_color;
-	float		ambient_ratio;
+	double		ambient_ratio;
 	bool		is_rendering;
 	t_color		background_color;
 	t_texture	skybox;
